Move split functions to ModelSplit.py
